+++ src/Components/Form/index.tsx
@@ -1,13 +1,15 @@
-import FormField from "../FormField";
+import CardNumberField from "../FormField/CardNumberField";
+import CardOwnerField from "../FormField/CardOwnerField";
+import CardValidityPeriodField from "../FormField/CardValidityPeriodField";
 import ErrorContextProvider from "./ErrorContextProvider";
 
 const Form = () => {
   return (
     <form>
       <ErrorContextProvider>
-        <FormField.CardNumberField />
-        <FormField.CardValidityPeriodField />
-        <FormField.CardOwnerField />
+        <CardNumberField />
+        <CardValidityPeriodField />
+        <CardOwnerField />
       </ErrorContextProvider>
     </form>
   );
